perf: 调整主题配置文件结构
