--- src/utils/theme/themeConfigs/chinese_red.js
@@ -1,17 +1,0 @@
-const config = {
-  chinese_red: {
-    '--color-primary': '#bd1a2d',
-    '--color-success': '#E98737',
-    '--color-info': '#f06f06',
-    '--color-warning': '#f82ed3',
-    '--color-danger': '#074bdc',
-    '--menu-bg': '#b4312b',
-    '--menu-text': '#fff1d9',
-    '--menu-hover': '#a22c27',
-    '--submenu-bg': '#902722',
-    '--submenu-hover': '#7e221e',
-    '--menu-active-left': 'rgb(241, 212, 174)'
-  }
-}
-
-export default config
