--- src/utils/theme/themeConfigs/deep_black.js
@@ -1,17 +1,0 @@
-const config = {
-  deep_black: {
-    '--color-primary': '#0576b9',
-    '--color-success': '#2cb57e',
-    '--color-info': '#47bce8',
-    '--color-warning': '#f5bd58',
-    '--color-danger': '#f56e50',
-    '--menu-bg': '#121e2d',
-    '--menu-text': '#8fa6bf',
-    '--menu-hover': '#101b29',
-    '--submenu-bg': '#0e1824',
-    '--submenu-hover': '#0d1520',
-    '--menu-active-left': '#0576b9'
-  }
-}
-
-export default config
