--- src/utils/theme/themeConfigs/default.js
@@ -1,17 +1,0 @@
-const config = {
-  default: {
-    '--color-primary': '#1ab394',
-    '--color-success': '#1c84c6',
-    '--color-info': '#23c6c8',
-    '--color-warning': '#f8ac59',
-    '--color-danger': '#ed5565',
-    '--menu-bg': '#2f4050',
-    '--menu-text': '#a7b1c2',
-    '--menu-hover': '#293846',
-    '--submenu-bg': '#1f2d3d',
-    '--submenu-hover': '#3b4c5b',
-    '--menu-active-left': '#1ab394'
-  }
-}
-
-export default config
