--- src/utils/theme/themeConfigs/tech_blue.js
@@ -1,17 +1,0 @@
-const config = {
-  tech_blue: {
-    '--color-primary': '#174779',
-    '--color-success': '#174779',
-    '--color-info': '#17a2b8',
-    '--color-warning': '#ffc107',
-    '--color-danger': '#dc3545',
-    '--menu-bg': '#174779',
-    '--menu-text': '#e9ecef',
-    '--menu-hover': '#123961',
-    '--submenu-bg': '#123961',
-    '--submenu-hover': '#103255',
-    '--menu-active-left': 'white'
-  }
-}
-
-export default config
